Initial version of rebaser.sh
The purpose of the rebaser is to be a slightly more advanced version of
an interactive, merge-preserving rebase. It supports not only rebasing
topic branches, keeping their merge structure intact, but also allows
for reordering commits, as well as integrating msysGit's 'merging
rebase' strategy.

To support all that, this introduces a couple of new "rebase commands"
by installing a temporary alias called ".r" that simply calls rebaser.sh
with one of the following commands: "merge", "start_merging_rebase" and
"cleanup". As a consequence, the edit scripts look very different from
current 'rebase -p' scripts.

Example:

	exec git .r start_merging_rebase "2fb434d"
	exec git update-ref refs/rewritten/onto HEAD

	exec git reset --hard refs/rewritten/onto
	pick 780e6c6 Add a ZipInputStream
	exec git update-ref refs/rewritten/780e6c6 HEAD

	exec git reset --hard refs/rewritten/onto
	exec git .r merge refs/rewritten/780e6c6 -C 2fb434d # Merge branch 'zip-input-stream' into HEAD

	exec git .r cleanup  780e6c6

Subsequent commits will improve this script further.

Since this developer designed git-rebase's --preserve-merges support in
a very unfortunate, awfully short-sighted manner, it was sadly not
possible to provide this functionality as a patch series improving
git-rebase itself, but it is rather an add-on script that circumvents
parts of git-rebase.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>